Manually corrected the web.config
Corrected the address for issuerMetadata -- where did it even come up with
that url?
Added the issuer element and referenced an appropriate binding for
communicating with the identity server.

Updated service reference again -- most of the values come through
correctly. No binding specified for the issuer causes the service request
to still fail.